feat(ui): DOMA-11759 refactor
