--- packages/ui/src/components/Input/textArea/textArea.less
@@ -1,52 +1,0 @@
-@import (reference) "@open-condo/ui/src/tokens/variables.less";
-@import (reference) "@open-condo/ui/src/components/style/mixins/typography";
-
-.condo-input {
-  &.condo-input-textarea {
-    height: auto;
-    padding: @condo-global-spacing-12;
-    border: none !important;
-  }
-}
-
-.condo-input-textarea-wrapper {
-  position: relative;
-  height: fit-content !important;
-  padding: 0 !important;
-}
-
-.condo-input-utils {
-  display: flex;
-  gap: @condo-global-spacing-8;
-  align-items: center;
-  max-height: @condo-global-spacing-20;
-}
-
-.condo-input-count {
-  color: @condo-global-color-gray-7;
-  font-size: @condo-global-font-size-text-s;
-  font-family: "Noto Sans Mono", monospace;
-}
-
-.condo-input-bottom-panel {
-  position: relative;
-  left: @condo-global-spacing-12;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: calc(100% - @condo-global-spacing-12 * 2);
-  max-height: @condo-global-spacing-12 + @condo-global-spacing-32;
-  padding: @condo-global-spacing-8 0 @condo-global-spacing-12;
-  background-color: @condo-global-color-white;
-
-  &-right {
-    display: flex;
-    gap: @condo-global-spacing-16;
-    align-items: center;
-    margin-left: auto;
-  }
-}
-
-.arrow-icon {
-  transform: rotate(-90deg);
-}
